Orbileads-backend/requirements.txt
Django>=5.1,<6.1
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
psycopg[binary]>=3.2,<4.0
python-dotenv>=1.0,<2.0
gunicorn>=23.0,<24.0
whitenoise>=6.8,<7.0
